+++ modules/bootstrapped/test/src-js/smithy4s/TimestampSpec.scala
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2021-2023 Disney Streaming
+ *  Copyright 2021-2024 Disney Streaming
  *
  *  Licensed under the Tomorrow Open Source Technology License, Version 1.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -214,6 +214,32 @@ class TimestampSpec() extends munit.FunSuite with munit.ScalaCheckSuite {
       val tsFromEpochMilli = Timestamp.fromEpochMilli(epochMilli)
       expect.same(tsEpochMilli, epochMilli)
       expect.same(tsFromEpochMilli, ts)
+    }
+  }
+
+  property("Truncate to milliseconds precision") {
+    forAll { (d: Date) =>
+      val epochMilli = d.valueOf().toLong
+      val ts = Timestamp.fromDate(d).truncateToMillis
+
+      val strippedDate = new Date(0)
+      strippedDate.setUTCMilliseconds(epochMilli)
+      val tsFromStrippedDate = Timestamp.fromDate(strippedDate)
+
+      expect.same(ts, tsFromStrippedDate)
+    }
+  }
+
+  property("Truncate to seconds precision") {
+    forAll { (d: Date) =>
+      val epochSecond = (d.valueOf() / 1000).toLong
+      val ts = Timestamp.fromDate(d).truncateToMillis
+
+      val strippedDate = new Date(0)
+      strippedDate.setUTCSeconds(epochSecond)
+      val tsFromStrippedDate = Timestamp.fromDate(strippedDate)
+
+      expect.same(ts, tsFromStrippedDate)
     }
   }
 }